SHARED-82748 Create Alpha OL Container for R&D and QA
Moved image related files into image subfolder. This will reduce the image size as the iso file is not in the image anymore.

Fixed an issue in the build script that did not delete all setup files from a previous run. Increased sleep to allow extraction of setup files to fully complete.